requirements.txt
openai-whisper
pyannote.audio
hypothesis
pytest
